Extract component to common directory
