+++ src/app/components/list/list.component.css
