+++ src/app/components/main/main.component.spec.ts
@@ -1,31 +1,31 @@
 /* tslint:disable:no-unused-variable */
 
 import { TestBed, async } from '@angular/core/testing';
-import { AppComponent } from './app.component';
+import { MainComponent } from './main.component';
 
-describe('AppComponent', () => {
+describe('MainComponent', () => {
     beforeEach(() => {
         TestBed.configureTestingModule({
             declarations: [
-                AppComponent
+                MainComponent
             ],
         });
     });
 
     it('should create the app', async(() => {
-        let fixture = TestBed.createComponent(AppComponent);
+        let fixture = TestBed.createComponent(MainComponent);
         let app = fixture.debugElement.componentInstance;
         expect(app).toBeTruthy();
     }));
 
     it(`should have as title 'app works!'`, async(() => {
-        let fixture = TestBed.createComponent(AppComponent);
+        let fixture = TestBed.createComponent(MainComponent);
         let app = fixture.debugElement.componentInstance;
         expect(app.title).toEqual('app works!');
     }));
 
     it('should render title in a h1 tag', async(() => {
-        let fixture = TestBed.createComponent(AppComponent);
+        let fixture = TestBed.createComponent(MainComponent);
         fixture.detectChanges();
         let compiled = fixture.debugElement.nativeElement;
         expect(compiled.querySelector('h1').textContent).toContain('app works!');
