+++ src/app/index.ts
@@ -1,2 +1,2 @@
-export * from './app.component';
-export * from './app.module';
+export * from './components/main/main.component';
+export * from './components/main/main.module';
